+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # datannur
 
+## unreleased
+
+- changed : description info with correct padding and width on mobile
+
 ## 0.9.4 (2024-11-12)
 
 - added : table column filter info popup